chore(core): organize submodules
