+++ packages/core/src/submodules/httpAuthSchemes/aws_sdk/index.ts
