+++ packages/core/src/submodules/httpAuthSchemes/index.ts
