+++ packages/core/src/submodules/httpAuthSchemes/utils/getDateHeader.ts
