+++ packages/core/src/submodules/httpAuthSchemes/utils/index.ts
