+++ packages/core/src/submodules/protocols/json/awsExpectUnion.spec.ts
