+++ packages/core/src/submodules/protocols/json/awsExpectUnion.ts
